Module cc::windows_registry

A helper module to looking for windows-specific tools:

On Windows host, probe the Windows Registry if needed;

On non-Windows host, check specified environment variables.

Enums

VsVers

A version of Visual Studio

Functions

find

Attempts to find a tool within an MSVC installation using the Windows registry as a point to search from.

find_tool

Similar to the find function above, this function will attempt the same operation (finding a MSVC tool in a local install) but instead returns a Tool which may be introspected.

find_vs_version

Find the most recent installed version of Visual Studio