cc::windows_registry
Function find

pub fn find(arch_or_target: &str, tool: &str) -> Option<Command>

Attempts to find a tool within an MSVC installation using the Windows registry as a point to search from.

The arch_or_target argument is the architecture or the Rust target triple that the tool should work for (e.g. compile or link for). The supported architecture names are:

"i586"

"i686"

"x86_64"

"arm"

"thumbv7a"

"aarch64"

"arm64ec"

The tool argument is the tool to find (e.g. cl.exe or link.exe).

This function will return None if the tool could not be found, or it will return Some(cmd) which represents a command that’s ready to execute the tool with the appropriate environment variables set.

Note that this function always returns None for non-MSVC targets (if a full target name was specified).