Function cc::windows_registry::find

pub fn find(target: &str, tool: &str) -> Option<Command>

Attempts to find a tool within an MSVC installation using the Windows registry as a point to search from.

The target argument is the target that the tool should work for (e.g. compile or link for) and the tool argument is the tool to find (e.g. cl.exe or link.exe).

This function will return None if the tool could not be found, or it will return Some(cmd) which represents a command that’s ready to execute the tool with the appropriate environment variables set.

Note that this function always returns None for non-MSVC targets.